+++ Eatery Blue/UI/Onboarding/OnboardingFeaturesViewController.swift
@@ -220,7 +220,7 @@ class OnboardingFeaturesViewController: UIViewController {
         if nextIndex == pages.count {
             // We've reached the last page
             
-            navigationController?.pushViewController(OldHomeModelController(), animated: true)
+            navigationController?.pushViewController(HomeViewController(), animated: true)
 
         } else {
             // Move to the next page
